% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot Patient Condition Trends}
\usage{
plotTrend(data, byMonthonth = FALSE, condition = NULL)
}
\arguments{
\item{data}{A data frame with columns: condition, year, month, count}

\item{condition}{Character, a specific condition to filter by. If NULL, plots all conditions (default: NULL), CASE SENSITIVE}

\item{byMonth}{Logical, if TRUE, plot by month; otherwise, plot by year (default)}
}
\value{
A ggplot2 object
}
\description{
This function takes the output from the \code{extractPatients} function as input and returns a plot
with the number of patients per year (default) or per month for each condition in the data.
}
\examples{
\dontrun{
connection <- DatabaseConnector::connect(Eunomia::getEunomiaConnectionDetails())
patient_data <- extractPatients(connection)
plotTrend(patient_data)
DatabaseConnector::disconnect(connection)
}
}
